pyproject.toml
[project]
name = "github-discussions-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "numpy>=2.2.4",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
]
